taph/requirements.txt
blinker~=1.8.2
boto3~=1.35.17
botocore~=1.35.17
cachelib~=0.13.0
cffi~=1.17.1
click~=8.1.7
cryptography~=43.0.1
Flask~=3.0.3
Flask-Session~=0.8.0
Flask-WTF~=1.2.1
gunicorn~=23.0.0
itsdangerous~=2.2.0
Jinja2~=3.1.4
jmespath~=1.0.1
MarkupSafe~=2.1.5
msgspec~=0.18.6
packaging~=24.1
pycparser~=2.22
PyJWT~=2.9.0
python-dateutil~=2.9.0.post0
redis~=5.2.0
s3transfer~=0.10.2
six~=1.16.0
urllib3~=2.2.2
valkey~=6.0.2
Werkzeug~=3.0.4
WTForms~=3.1.2